refactor(utils/moreurls): create separate packages for parameter type and parameters type (to have proper private/public support)
