+++ src/search/engines/bing/search.go
@@ -15,6 +15,7 @@ import (
 	"github.com/hearchco/agent/src/search/scraper/parse"
 	"github.com/hearchco/agent/src/utils/anonymize"
 	"github.com/hearchco/agent/src/utils/moreurls"
+	"github.com/hearchco/agent/src/utils/moreurls/parameters"
 )
 
 func (se Engine) Search(query string, opts options.Options, resChan chan result.ResultScraped) ([]error, bool) {
@@ -81,13 +82,13 @@ func (se Engine) Search(query string, opts options.Options, resChan chan result.
 		ctx.Put("page", strconv.Itoa(i))
 
 		// Build the parameters.
-		params := moreurls.NewParams(
+		params := parameters.NewParams(
 			paramQueryK, query,
 			paramLocaleK, paramLocaleV,
 			paramLocaleSecK, paramLocaleSecV,
 		)
 		if pageNum0 > 0 {
-			params = moreurls.NewParams(
+			params = parameters.NewParams(
 				paramQueryK, query,
 				paramPageK, strconv.Itoa(pageNum0*10+1),
 				paramLocaleK, paramLocaleV,
