+++ src/search/engines/brave/search.go
@@ -14,6 +14,7 @@ import (
 	"github.com/hearchco/agent/src/search/scraper/parse"
 	"github.com/hearchco/agent/src/utils/anonymize"
 	"github.com/hearchco/agent/src/utils/moreurls"
+	"github.com/hearchco/agent/src/utils/moreurls/parameters"
 )
 
 func (se Engine) Search(query string, opts options.Options, resChan chan result.ResultScraped) ([]error, bool) {
@@ -69,12 +70,12 @@ func (se Engine) Search(query string, opts options.Options, resChan chan result.
 		ctx.Put("page", strconv.Itoa(i))
 
 		// Build the parameters.
-		params := moreurls.NewParams(
+		params := parameters.NewParams(
 			paramQueryK, query,
 			paramSourceK, paramSourceV,
 		)
 		if pageNum0 > 0 {
-			params = moreurls.NewParams(
+			params = parameters.NewParams(
 				paramQueryK, query,
 				paramPageK, strconv.Itoa(pageNum0),
 				paramSpellcheckK, paramSpellcheckV,
