+++ src/search/engines/duckduckgo/suggest.go
@@ -11,6 +11,7 @@ import (
 	"github.com/hearchco/agent/src/search/scraper"
 	"github.com/hearchco/agent/src/utils/anonymize"
 	"github.com/hearchco/agent/src/utils/moreurls"
+	"github.com/hearchco/agent/src/utils/moreurls/parameters"
 )
 
 func (se Engine) Suggest(query string, options options.Options, sugChan chan result.SuggestionScraped) ([]error, bool) {
@@ -48,7 +49,7 @@ func (se Engine) Suggest(query string, options options.Options, sugChan chan res
 	ctx := colly.NewContext()
 
 	// Build the parameters.
-	params := moreurls.NewParams(
+	params := parameters.NewParams(
 		paramQueryK, query,
 		sugParamTypeK, sugParamTypeV,
 	)
