+++ src/search/engines/googlescholar/search.go
@@ -14,6 +14,7 @@ import (
 	"github.com/hearchco/agent/src/search/scraper/parse"
 	"github.com/hearchco/agent/src/utils/anonymize"
 	"github.com/hearchco/agent/src/utils/moreurls"
+	"github.com/hearchco/agent/src/utils/moreurls/parameters"
 )
 
 func (se Engine) Search(query string, opts options.Options, resChan chan result.ResultScraped) ([]error, bool) {
@@ -72,7 +73,7 @@ func (se Engine) Search(query string, opts options.Options, resChan chan result.
 		ctx.Put("page", strconv.Itoa(i))
 
 		// Build the parameters.
-		params := moreurls.NewParams(
+		params := parameters.NewParams(
 			paramQueryK, query,
 			paramFilterK, paramFilterV,
 			paramLocaleK, paramLocaleV,
@@ -80,7 +81,7 @@ func (se Engine) Search(query string, opts options.Options, resChan chan result.
 			paramSafeSearchK, paramSafeSearchV,
 		)
 		if pageNum0 > 0 {
-			params = moreurls.NewParams(
+			params = parameters.NewParams(
 				paramQueryK, query,
 				paramFilterK, paramFilterV,
 				paramPageK, strconv.Itoa(pageNum0*10),
